refactor: move plugin-helpers into ensnode app
