+++ apps/ensnode/ponder.config.ts
@@ -1,4 +1,4 @@
-import { type MergedTypes, getActivePlugins } from "ensnode-common/plugin-helpers";
+import { type MergedTypes, getActivePlugins } from "./src/lib/plugin-helpers";
 import { deepMergeRecursive } from "./src/lib/ponder-helpers";
 import * as baseEthPlugin from "./src/plugins/base.eth/ponder.config";
 import * as ethPlugin from "./src/plugins/eth/ponder.config";
